bug fix for graalvm support #241

+++ core/src/main/java/com/alibaba/fastjson2/reader/FieldReaderInt8ValueMethod.java
@@ -47,6 +47,6 @@ final class FieldReaderInt8ValueMethod<T> extends FieldReaderObjectMethod<T> {
 
     @Override
     public Object readFieldValue(JSONReader jsonReader) {
-        return jsonReader.readInt32Value();
+        return (byte) jsonReader.readInt32Value();
     }
 }
